apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI backend for the Revio"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "chromadb>=0.6.0",
    "sqlalchemy>=2.0.46",
    "alembic>=1.18.4",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "google-genai>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "livekit-api>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
]
voice = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-google>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -rN"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
